requirements.txt
python-dotenv
pandas 
numpy
pymongo==3.11
certifi
pymongo[srv]==3.11


-e.